cart: mark the choosen attribute with highest Gini gain to be skipped.
Unlike ID3, where the attribute is removed once it has been selected as
split criteria, the CART algorithm allow the attribute to be computed
again except in their children node.

When the gain has been computed, the childs of splitted node MUST not
contain the same attribute as the current split attribute, but the
grandchild process (node) can have the same attribute with their
grand-parent.