cart: fix splitting tree.
If max gain value is zero, use the majority class as node target value
and terminate the split function.

This is the case where all attribute value is only zero or equal, which
make splitting data result as empty on the left side and the same data in
the right side of split, which subsequently cause infinite recursive when
calling split function the right side of split.