cart: fix on setting columns flag.
Turn out using "key, val" in range, the val is a copy of element, and
when we set their attribute the original object is not changed. This
cause flag on columns is not changed which slightly effect the
computation.

Also add function to classify a single row.

On SelectRandomFeature, check if number of random feature is greater or
equal with number of dataset features. If its true, no random feature
will be used but all features.